requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
gitpython==3.1.43
torch==2.3.0
torch-geometric==2.5.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
anthropic==0.28.0
transformers==4.41.2
numpy==1.26.4
networkx==3.3
tree-sitter==0.21.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.2
pydantic-settings==2.3.0
tqdm==4.66.4
scikit-learn==1.5.0